// src/enums/cause-of-failure.enum.ts
export enum CausesOfFailureEnum {
  INSUFFICIENT_COFFEE = 'Insufficient coffee',
  OVERSUFFICIENT_COFFEE = 'Oversufficient coffee',
  INSUFFICIENT_MONEY = 'Insufficient money',
  OVERSUFFICIENT_MONEY = 'Oversufficient money',
  INSUFFICIENT_PERSONNEL = 'Insufficient personnel',
  OVERSUFFICIENT_PERSONNEL = 'Oversufficient personnel',
  INSUFFICIENT_CUSTOMERS = 'Insufficient customers',
  OVERSUFFICIENT_CUSTOMERS = 'Oversufficient customers',
}